fixed some view after review
