+++ api_yamdb/api/urls.py
@@ -1,31 +1,6 @@
 from django.urls import include, path
-from rest_framework.routers import DefaultRouter
-
-from .views import (CategoryViewSet, CommentViewSet, CurrentUser, GenreViewSet,
-                    ReviewViewSet, TitleViewSet, UserViewSet, signup, token)
-
-app_name = 'api'
-router = DefaultRouter()
-
-router.register('users', UserViewSet, basename='users')
-router.register('categories', CategoryViewSet, basename='сategories')
-router.register('genres', GenreViewSet, basename='genres')
-router.register('titles', TitleViewSet, basename='titles')
-router.register(
-    r'titles/(?P<title_id>\d+)/reviews',
-    ReviewViewSet,
-    basename='reviews'
-)
-router.register(
-    r'titles/(?P<title_id>\d+)/reviews/(?P<review_id>\d+)/comments',
-    CommentViewSet,
-    basename='comments'
-)
-
+from .v1 import urls as v1_urls
 
 urlpatterns = [
-    path('v1/auth/signup/', signup, name="signup"),
-    path('v1/auth/token/', token, name="token"),
-    path('v1/users/me/', CurrentUser.as_view(), name="users/me"),
-    path('v1/', include(router.urls)),
+    path('v1/', include(v1_urls))
 ]
