+++ api_yamdb/api/v1/filters.py
@@ -1,0 +1,13 @@
+from reviews.models import Title
+from django_filters import CharFilter, FilterSet, NumberFilter
+
+
+class TitleFilterSet(FilterSet):
+    category = CharFilter(field_name='category__slug', lookup_expr='icontains')
+    genre = CharFilter(field_name='genre__slug', lookup_expr='icontains')
+    name = CharFilter(field_name='name', lookup_expr='icontains')
+    year = NumberFilter(field_name='year', lookup_expr='icontains')
+
+    class Meta:
+        model = Title
+        fields = ('category', 'genre', 'name', 'year')
