+++ api_yamdb/api/v1/mixins.py
@@ -1,0 +1,10 @@
+from rest_framework import mixins, viewsets
+
+
+class CustomMixin(
+    mixins.ListModelMixin,
+    mixins.CreateModelMixin,
+    mixins.DestroyModelMixin,
+    viewsets.GenericViewSet
+):
+    pass
